+++ primus_turbo/pytorch/ops/attention/__init__.py
@@ -1,2 +1,2 @@
 from .flash_attn_interface import flash_attn_fp8_func, flash_attn_func
-from .usp import flash_attn_fp8_usp_func, flash_attn_usp_func
+from .flash_attn_usp_interface import flash_attn_fp8_usp_func, flash_attn_usp_func
