--- primus_turbo/pytorch/ops/attention/usp/__init__.py
@@ -1,1 +1,0 @@
-from .flash_attn_usp_interface import flash_attn_fp8_usp_func, flash_attn_usp_func
